Adding auth to user, administrator requests

+++ api/src/components/administrator/administrator.service.ts
@@ -10,10 +10,7 @@ class AdministratorAdapterOptions implements IModelAdapterOptions { }
 
 export default class AdministratorService extends BaseService<Administrator> {
 
-    async adaptToModel(
-        data: any,
-        options: Partial<AdministratorAdapterOptions>
-    ): Promise<Administrator> {
+    async adaptToModel(data: any, opt: Partial<AdministratorAdapterOptions>): Promise<Administrator> {
         const item: Administrator = new Administrator();
         item.administratorId = Number(data?.administrator_id);
         item.email = data?.administrator_email;
@@ -22,20 +19,15 @@ export default class AdministratorService extends BaseService<Administrator> {
         return item;
     }
 
-    public async getAll(
-        options: Partial<AdministratorAdapterOptions> = { loadChildren: true }
-    ): Promise<Administrator[]> {
+    async getAll(options: Partial<AdministratorAdapterOptions> = { loadChildren: true }): Promise<Administrator[]> {
         return this.getAllFromTable<AdministratorAdapterOptions>("administrator", options);
     }
 
-    public async getById(
-        adminId: number,
-        options: Partial<AdministratorAdapterOptions> = { loadChildren: true }
-    ): Promise<Administrator | null> {
+    async getById(adminId: number, options: Partial<AdministratorAdapterOptions> = { loadChildren: true }): Promise<Administrator | null> {
         return super.getByIdFromTable<AdministratorAdapterOptions>("administrator", adminId, options);
     }
 
-    public async getByEmail(email: string, options: Partial<AdministratorAdapterOptions> = {}): Promise<Administrator | null> {
+    async getByEmail(email: string, options: Partial<AdministratorAdapterOptions> = {}): Promise<Administrator | null> {
         const result = await super.getByFieldIdFromTable<AdministratorAdapterOptions>('administrator', 'administrator_email', email, options);
         if (!Array.isArray(result) || result.length === 0) {
             return null;
@@ -43,18 +35,21 @@ export default class AdministratorService extends BaseService<Administrator> {
         return result[0];
     }
 
-    public async create(data: ICreateAdministrator): Promise<Administrator | IErrorResponse> {
+    async create(data: ICreateAdministrator): Promise<Administrator | IErrorResponse> {
         return new Promise<Administrator | IErrorResponse>((result) => {
-            const sql: string = `
+
+            const passwordHash = bcypt.hashSync(data.password, 12);
+
+            this.db.execute(`
                 INSERT
                     administrator
                 SET
                     administrator_email = ?,
-                    password_hash = ?;`;
-
-            const passwordHash = bcypt.hashSync(data.password, 12);
-
-            this.db.execute(sql, [data.email, passwordHash])
+                    password_hash = ?;`,
+                [
+                    data.email,
+                    passwordHash
+                ])
                 .then(async res => {
                     const resultData: any = { ...res };
                     const newAdminId: number = Number(resultData[0]?.insertId);
@@ -69,20 +64,23 @@ export default class AdministratorService extends BaseService<Administrator> {
         });
     }
 
-    public async update(adminId: number, data: IUpdateAdministrator): Promise<Administrator | IErrorResponse> {
+    async update(adminId: number, data: IUpdateAdministrator): Promise<Administrator | IErrorResponse> {
         return new Promise<Administrator | IErrorResponse>((result) => {
-            const sql: string = `
+
+            const passwordHash = bcypt.hashSync(data.password, 12);
+
+            this.db.execute(`
                 UPDATE
                     administrator
                 SET
                     password_hash = ?
                 WHERE
-                    administrator_id = ?;`;
-
-            const passwordHash = bcypt.hashSync(data.password, 12);
-
-            this.db.execute(sql, [passwordHash, adminId])
-                .then(async res => {
+                    administrator_id = ?;`,
+                [
+                    passwordHash,
+                    adminId
+                ])
+                .then(async _ => {
                     result(await this.getById(adminId, { loadChildren: true }));
                 })
                 .catch(err => {
@@ -94,15 +92,9 @@ export default class AdministratorService extends BaseService<Administrator> {
         });
     }
 
-    public async delete(adminId: number): Promise<IErrorResponse> {
+    async delete(adminId: number): Promise<IErrorResponse> {
         return new Promise<IErrorResponse>((result) => {
-            const sql: string = `
-                DELETE FROM
-                    administrator
-                WHERE
-                    administrator_id = ?;`;
-
-            this.db.execute(sql, [adminId])
+            this.db.execute(`DELETE FROM administrator WHERE administrator_id = ?;`, [adminId])
                 .then(async res => {
                     const data: any = res;
                     result({
@@ -118,5 +110,4 @@ export default class AdministratorService extends BaseService<Administrator> {
                 });
         });
     }
-
 }
